+++ apps/web/app/(app)/(organization)/settings/billing/page.tsx
@@ -1,10 +1,21 @@
+import { BreadcrumbSetter } from "@workspace/ui/components/setter-breadcrumb";
 import { ReactElement } from "react";
 
 const OrganizationBilling = (): ReactElement => {
   return (
-    <div>
-      <h1>Organization Billing</h1>
-    </div>
+    <>
+      <BreadcrumbSetter items={
+        [
+          { label: "Organization", href: "/" },
+          { label: "Settings", href: "/settings" },
+          { label: "Billing" }
+        ]
+      } />
+
+      <div>
+        <h1>Organization Billing</h1>
+      </div>
+    </>
   );
 }
